Fixed create_server
Creating a server with a template is different now. You have to POST guilds/templates/${guild_template_code}.
I also added support for server icons. This code now only ever creates servers based on a template. But if no template is specified it just uses the default one, so it's okay.